Refactor: PageInfoCard 공용컴포넌트화
